Windows 体验收尾：控制台 UTF-8 输出 + 平台能力检测归一
- cli.py：Windows 上强制 stdout/stderr 为 UTF-8（reconfigure）并将控制台
  输出码页切至 65001。此前 harness 的 UTF-8 产物（中文报告/事件）经 cp936
  管道或控制台输出时乱码，含 ✓ 等字符时甚至直接 UnicodeEncodeError 崩溃。
- utils/platform_caps.py（新增）：SECURE_DIRFD 能量检测与
  validate_no_symlink_chain 的唯一权威定义，从 control_bus 上移；control_bus
  保留私有别名，全部既有引用（manager/cli/tests）零改动。
- utils/platform_shell.py：IS_WINDOWS 改为引用 platform_caps，消除重复定义。

测试：全量 402 passed / 4 skipped 保持不变。
